pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plane-agent"
version = "0.1.2"
description = "Plane MCP Agent"
readme = "README.md"
authors = [{ name = "Audel Rouhi", email = "knucklessg1@gmail.com" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.1",
    "httpx>=0.27.0",
    "pydantic>=2.8.2",
    "fastmcp>=3.0.0b1",
    "eunomia-mcp>=0.3.10",
    "pydantic-ai-slim[fastmcp,openai,anthropic,groq,mistral,google,huggingface,a2a,ag-ui,web]>=1.58.0",
    "pydantic-ai-skills>=v0.4.0",
    "fastapi>=0.128.0"
]

[project.scripts]
plane-mcp = "plane_agent.mcp:main"
plane-agent = "plane_agent.agent:main"

[tool.setuptools.package-data]
plane_agent = ["mcp_config.json", "skills/**"]

[tool.setuptools.packages.find]
where = ["."]
